feat(admin): migrate to modular admin structure + fix OrderItem
  Admin refactoring:
- shop/admin.py → shop/admin/category_admin.py, product_admin.py, etc.
- shop/admin/__init__.py for auto-imports
- OrderAdmin + OrderItemInline (line items visible in orders)

  Bug fixes
- OrderItem.quantity: PositiveBigIntegerField → PositiveIntegerField
- Migration 0007 applied successfully

+++ shop/models/order_item.py
@@ -26,7 +26,7 @@ class OrderItem(models.Model):
         verbose_name=_("Product")
     )
 
-    quantity = models.PositiveBigIntegerField(
+    quantity = models.PositiveIntegerField(
         validators=[
             MinValueValidator(MIN_QUANTITY),
             MaxValueValidator(MAX_QUANTITY)
